change(ServerSide): Remove server side rendering code (#24)
* Remove server-side rendering code for now

It never really worked correctly. In order to get it working
correctly, some fiddliness would have to be done to get react stdio
working correctly. Then, in order for the compiled react components to
be able to be rendered, the compiled render path would have to be
inside node_modules. I don't think that is a great solution, and so to
reduce confusion in the package, let's remove it and make it
experimental instead.

* update mod

* update npm packages

* Update README and add Phoenix 1.2-specific README

* Remove old server-side tests

* correct version numbers in readme

* Correct paths in readme

* Add link to server-side rendering release 0.4.3